Fork release, unit tests to add for java 8 date TestParsing.java:128 and TestFormats.java:20 failing

+++ src/test/java/com/solab/iso8583/parse/CustomBinCodecs.java
@@ -84,7 +84,7 @@ public class CustomBinCodecs {
         }
         //Test parsing
         tmpl = mfact.parseMessage(buf, 0);
-        Assert.assertEquals(1234567890l, tmpl.getObjectValue(2));
+        Assert.assertEquals(1234567890l, (long)tmpl.getObjectValue(2));
         Assert.assertEquals(b29, tmpl.getObjectValue(3));
     }
